added response_model for delete_game
