+++ civ_vi_webhook/api/info_endpoints.py
@@ -1,35 +1,10 @@
 from fastapi import HTTPException, APIRouter, Query
 from typing import Optional
 
-from ..dependencies import load_most_recent_games
-from ..models import information_models, games
+from ..dependencies import load_most_recent_games, dict_to_game_model
+from ..models import information_models
 
 router = APIRouter(tags=['Information Endpoints'])
-
-
-def dict_to_game_model(dictionary: dict) -> games.Game:
-    """Take in a dictionary of a game and turn it into a Game model.
-
-    Example dict:
-
-    {"Eric's Barbarian Clash Game": {'player_name': 'Eric', 'turn_number': 300,
-    'time_stamp': {'year': 2022, 'month': 7, 'day': 21, 'hour': 20, 'minute': 33,
-    'second': 28}}}
-
-    """
-    game_name = list(dictionary.keys())
-    game_name = game_name[0]
-    time_stamp = games.TimeStamp(year=dictionary[game_name]['time_stamp']['year'],
-                                 month=dictionary[game_name]['time_stamp']['month'],
-                                 day=dictionary[game_name]['time_stamp']['day'],
-                                 hour=dictionary[game_name]['time_stamp']['hour'],
-                                 minute=dictionary[game_name]['time_stamp']['minute'],
-                                 second=dictionary[game_name]['time_stamp']['second'])
-    game_info = games.GameInfo(player_name=dictionary[game_name]['player_name'],
-                               turn_number=dictionary[game_name]['turn_number'],
-                               time_stamp=time_stamp)
-    game = games.Game(game_name=game_name, game_info=game_info)
-    return game
 
 
 @router.get('/current_games', response_model=information_models.CurrentGames)
