+++ civ_vi_webhook/models/action_models.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+from .games import Game
+
+
+class DeletedGame(BaseModel):
+    deleted_game: Game
